Creacion de relacion entre Reserva y una lista de clientes integrantes. Esta será usada para crear los detalleComprobante de cada uno. Ademas se crean funcion para agregar o quitar integrantes de reserva

+++ AppKartingRM/src/main/java/com/kartingRM/AppKartingRM/entities/ComprobanteEntity.java
@@ -20,7 +20,7 @@ public class ComprobanteEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long idComprobante;
     private boolean pagado;
-    private int total;
+    private double total;
 
     @OneToOne
     @JoinColumn(name = "id_reserva")
